+++ benchmarks/bench_plot_svd.py
@@ -8,11 +8,11 @@ import numpy as np
 from collections import defaultdict
 
 from scipy.linalg import svd
-from sklearn.utils.extmath import fast_svd
+from sklearn.utils.extmath import randomized_svd
 from sklearn.datasets.samples_generator import make_low_rank_matrix
 
 
-def compute_bench(samples_range, features_range, q=3, rank=50):
+def compute_bench(samples_range, features_range, n_iterations=3, rank=50):
 
     it = 0
 
@@ -36,16 +36,19 @@ def compute_bench(samples_range, features_range, q=3, rank=50):
             results['scipy svd'].append(time() - tstart)
 
             gc.collect()
-            print "benching scikit-learn fast_svd: q=0"
+            print "benching scikit-learn randomized_svd: n_iterations=0"
             tstart = time()
-            fast_svd(X, rank, q=0)
-            results['scikit-learn fast_svd (q=0)'].append(time() - tstart)
+            randomized_svd(X, rank, n_iterations=0)
+            results['scikit-learn randomized_svd (n_iterations=0)'].append(
+                time() - tstart)
 
             gc.collect()
-            print "benching scikit-learn fast_svd: q=%d " % q
+            print ("benching scikit-learn randomized_svd: n_iterations=%d "
+                   % n_iterations)
             tstart = time()
-            fast_svd(X, rank, q=q)
-            results['scikit-learn fast_svd (q=%d)' % q].append(time() - tstart)
+            randomized_svd(X, rank, n_iterations=n_iterations)
+            results['scikit-learn randomized_svd (n_iterations=%d)'
+                    % n_iterations].append(time() - tstart)
 
     return results